RESET
I (Lewis) that the best way to continue with this project is to start
over. The previous code was good for experimenting, but it is poorly
designed. This will allow incremental development instead of biting off
everything at once (i.e. begin with the quizzing UI, then slowly add
databases, etc). Of course, the previous code is still visible in
previous commits if we want to reference them.