Quiz UI: Dynamic HTML
The HTML of the quiz UI is modified based on the parameters of certain
state variables within the javascript of quiz.php (renamed from
quizOfficial.php). The load_quiz() function sets the inner HTML of the
teams div based on these variables. As part of this, the bulk of the
HTML is now stored and returned from a separate PHP file
(controller_get_quiz_teams_UI.php).

This commit is the precursor to changing the teams based on the lineups
menu.