Quiz UI: Move "dynamic" HTML from PHP to JS
Move the code that gets the teams' HTML from a PHP file to a JS file.
This is done so that the quizzing UI may be entirely static and thus
wouldn't have to be hosted on a server. (For now, the HTML content is
still in PHP files, so as it is this repo is not quite ready to be
completely statically hosted.)